Develop add lambdas

+++ dags/lambda_codes/ingest/parser/demo_parser.py
@@ -74,4 +74,7 @@ def delete_local_tmp(tempdirectory):
     print('deleting files: ',files_to_delete)
     for single_file in files_to_delete:
         filepath = os.path.join(tempdirectory, single_file)
-        os.remove(filepath)
+        if 'go' in filepath:
+            pass
+        else:
+            os.remove(filepath)
